// client/model/virtualServer.ts
/**
 * Draas
 * Draas OAS Documentation
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PoliciesReference } from './policiesReference';
import { ProfilesReference } from './profilesReference';
import { PoolReference } from './poolReference';
import { CertsReference } from './certsReference';
import { VirtualServerStatEntries } from './virtualServerStatEntries';


export interface VirtualServer { 
    readonly name: string;
    readonly description: string;
    readonly partition: string;
    readonly enabled: boolean;
    readonly addressStatus: string;
    readonly connectionLimit: number;
    readonly ipProtocol: string;
    readonly destination: string;
    readonly mask: string;
    readonly poolReference: PoolReference;
    readonly profilesReference: ProfilesReference;
    readonly policiesReference: PoliciesReference;
    readonly certsReference: Array<CertsReference>;
    readonly stats: VirtualServerStatEntries;
}